Fixed compile errors and warnings on MSVC

+++ modules/custom_operations/user_ie_extensions/grid_sample.cpp
@@ -33,31 +33,31 @@ bool GridSample::evaluate(ov::TensorVector& outputs, const ov::TensorVector& inp
     std::vector<size_t> inpDims = inputs[0].get_shape();
     std::vector<size_t> outDims = outputs[0].get_shape();
 
-    const int batch     = outDims[0];
-    const int channels  = outDims[1];
-    const int height    = outDims[2];
-    const int width     = outDims[3];
-    const int inpHeight = inpDims[2];
-    const int inpWidth  = inpDims[3];
-    const int inpPlane = inpHeight * inpWidth;
-    const int outPlane = height * width;
+    const size_t batch     = outDims[0];
+    const size_t channels  = outDims[1];
+    const size_t height    = outDims[2];
+    const size_t width     = outDims[3];
+    const size_t inpHeight = inpDims[2];
+    const size_t inpWidth  = inpDims[3];
+    const size_t inpPlane  = inpHeight * inpWidth;
+    const size_t outPlane  = height * width;
 
     std::vector<float> zerosPlane(inpDims[1] * inpDims[2] * inpDims[3], 0);
     float* zeros = zerosPlane.data();
 
-    ov::parallel_for(batch, [&](int d) {
+    ov::parallel_for(batch, [&](size_t d) {
         const float* inp  = inpData + d * channels * inpPlane;
         const float* grid = gridData + d * outPlane * 2;
-        for (int y = 0; y < height; ++y) {
-            for (int x = 0; x < width; ++x) {
-                int offset = y * width + x;
+        for (size_t y = 0; y < height; ++y) {
+            for (size_t x = 0; x < width; ++x) {
+                size_t offset = y * width + x;
 
                 float input_x = 0.5f * (grid[offset * 2] + 1) * (inpWidth - 1);
-                int x0 = std::floor(input_x);
+                int x0 = static_cast<int>(std::floor(input_x));
                 int x1 = x0 + 1;
 
                 float input_y = 0.5f * (grid[offset * 2 + 1] + 1) * (inpHeight - 1);
-                int y0 = std::floor(input_y);
+                int y0 = static_cast<int>(std::floor(input_y));
                 int y1 = y0 + 1;
 
                 const float* inp_row0 = (0 <= y0 && y0 < inpHeight) ? inp + y0 * inpWidth : zeros;
